Creación y manejo de clase Categoria

+++ subastas-online-tpi/src/main/java/com/subastas/tpi/dto/response/UsuarioResponseDTO.java
@@ -1,11 +1,10 @@
 package com.subastas.tpi.dto.response;
 
 public record UsuarioResponseDTO(
-    
-    Integer id,
-    String nombre,
-    String apellido,
-    String email,
-    String telefono,
-    Boolean activo
+        Long id,
+        String nombre,
+        String apellido,
+        String email,
+        String telefono,
+        Boolean activo
 ) {}
